Fix: Use Sun - Sat weeks for NALD
WATER-1632

Changes week calculations to us weeks that start on Sundays.

Sets the global moment locale to en-gb, but then adds a nald-dates
module for calculating the weeks, where a local locale of `en` is used
which treats weeks as running Sun - Sat.

Updates any existing code that used `startOf` or `endOf` and routes
through nald-dates.