Feat: Adds bulk returns upload job
WATER-1934

Adds the PG Boss Job that can be created with an event id, that in turn
will attempt to persist any returns from an XML that have already been
validated.

The result of the save attempt will be written back to the event.

Also updates some deps based on dependabot suggestions.