Feat: Map XML to JSON using return model shape
WATER-1935

Updates the XML to JSON mapping to align with the shape of the return
model used when entering the values in the UI for a single return.

This means in a different job, the data can be saved using the same
endpoint.

This changes the xml-to-json-mapping module to have to reach out to
downstream services to get more data to fulfil the interface of the
return model.