Feat: Adds add internal endpoint
WATER-2228

Adds a new resource at water/1.0/user/internal that will create a new
internal user and assign the appropriate roles and groups based on a
permission key that can be expanded into a set of groups and/or roles.

Updates to the latest water-abstraction-helpers to get the newly added
PUT method of the service-request.