Feat: Adds transactions to the invoice API
WATER-2414

Provides routes at

`/water/1.0/billing/batches/{batchId}/invoices/{invoiceId}`

and

`/water/1.0/billing/batches/{batchId}/invoices`

that get invoice data with transactions via the charge module.

Because the charge-module is not live at the moment the API is tolerant
of the requirement for the VPN access and returns an empty array of
transactions if none are available.

Updates the models to include `toJSON` to serialize the desired
properties.

Updates the Invoice model to provide a function that will get a totals
object.