PR feedback
Introduces a Model base class that takes on the `toJSON` functionalility
of remove the underscores from the keys that are output when serializing
to a string.

Adds ChargeModuleTransaction to wrap some of the properties from the
transactions response from the charging module api.

Changes the transaction service to return ChargeModuleTransaction
objects instead of partially/incorrect Transaction objects.

Moves more validation code to the validators module.