Sort the results of all possible cycles
In the UI we want to display the lastest (current) year first followed by the rest in descending order. Rather than complicate the UI code with logic to ensure this we add it to the API endpoint as that's where the data crunching is taking place.

We can add tests to ensure order is always as expected later.