Attempt to identify the current year in results
The ticket asks us to display a `Current year` hint text in the UI. After speaking to Stephan we remembered the existing behaviour has logic that attempts to do this.

```
const mostRecentCycle = cycles.find(cycle => {
    const isDueDatePast = moment(cycle.dueDate).isBefore(moment(refDate), 'day');
    const isSeasonMatch = cycle.isSummer === isSummer;
    return isDueDatePast && isSeasonMatch;
  });
```
> https://github.com/DEFRA/water-abstraction-ui/blob/17574c117cb729531f5288bc9fcc5eab247225dd/src/internal/modules/billing/lib/batch-financial-year.js#L14

This change attempts to bring that into our results but it's not working.

What we are now sending back is all years are marked as `current` because meet this definition. We need to get a better understanding of what defines 'current year'.