Attempt to solve lint issue in migrations
The issue is a number of variables are declared, in some cases assigned to but never appear to be used anywhere. standardjs has picked them up which causes linting to fail.

This is a 'on-blast' attempt to see if they are actually needed. Done with find & replace because of the number of fails there is a good chance this change will crash in 🔥