Use standard js for linting (#1696)
https://github.com/DEFRA/water-abstraction-team/issues/9

[standardjs](https://standardjs.com/) is the [standard linter](https://github.com/DEFRA/software-development-standards/blob/master/standards/javascript_standards.md) for Javascript in Defra. However, it seems the previous team took exception to the not using semi-colons so instead adopted a modified version using eslint.

The new team is used to working on Defra projects that follow standardjs and would rather stick to it on this service.

So, we are working through the projects, switching the linter used from eslint to standardjs and then correcting any issues we find.

** Notes

- Fix sonarcloud issue

> Ternary operators should not be nested
>
> Just because you _can_ do something, doesn't mean you should, and that's the case with nested ternary operations. Nesting ternary operators results in the kind of code that may seem clear as day when you write it, but six months later will leave maintainers (or worse - future you) scratching their heads and cursing.
>
> Instead, err on the side of clarity, and use another line to express the nested operation as a separate statement.

Pretty much sums up our experience with this code base! Though in this case, we think it's a ternary that has been split across lines due to its complexity rather than a nested one. That said, it was primed for an attempt to clean it up.