Fix multi year 2pt bill runs for ALCS
https://eaflood.atlassian.net/browse/WATER-3679

Since merging in [SROC annual billing -- feat/water-3419](https://github.com/DEFRA/water-abstraction-service/pull/1698) our changes for [Add simplified endpoints to retrieve the data relating to batches for a financial year](https://github.com/DEFRA/water-abstraction-service/pull/1695) have stopped working.

After checking we can see it's because the SROC billing changes put back the change to hard code the `toFinancialYearEnd` to be 2022 for everything but annual bill runs.

We need the create bill run process to accept and use the year selected during the 2PT bill runs in the UI. But we also need to protect against trying to run an ALCS (pre-sroc) bill run for anything later than 2023 (because that's impossible!)

This change applies a fix to handle both cases.