Remove unused dev dependency
Again, we run `npx depcheck` and remove the unneeded dev dependency it finds. Note that it claims `db-migrate-pg` isn't needed and we cannot find a reference to it, but the docs for `db-migrate` do state that it's needed. We may do further checks to see if this is the case but for now we leave it.