Fix non-prod delete all bill runs function
https://eaflood.atlassian.net/browse/WATER-3712

There is a feature when the app is running in non-production mode which enables from the UI the ability to delete all bill runs.

It relies on an end point in the service which is enabled via a feature toggle. Unfortunately, it looks like we broke it when we were tidying up the modules in `src/modules/billing/routes/batches.js` as part of [Fix module.exports](https://github.com/DEFRA/water-abstraction-team/issues/28).

This change fixes the issue and makes the endpoint accessible again.