Get tests passing 1
The tests seem to rely on env vars we could just as easily mock or default. In fact, some tests do that. But we're still getting failures which indicates we still need to set them.

Which ones specifically is not clear. So, we started with our minimum set and then we'll keep adding them back in until all tests are passing. We'll then try and see if we can replace real values for mock ones.

Hopefully at some point we'll have the tests passing with a reduced set of env vars being set.