Testing if this avoids cpu-features
The db-migrate package has a dependency on ssh2, which has an optional dependency on https://github.com/mscdex/cpu-features. In order to work cpu-features needs to be compiled each time npm ci is run (because it has to be built specific to the OS it's running on).

However, ssh2 in its docs states

> (Optional) cpu-features is set as an optional package dependency (you do not need to install it explicitly/separately from ssh2) that will be automatically built and used if possible. See the project's documentation for its own requirements.

We're just trialling if we provide `--no-optional` to npm ci whether this package gets skipped as we don't depend on using ssh to run migrations against our db (at least we think we don't!)