Remove snyk exception config (#1719)
https://github.com/DEFRA/water-abstraction-team/issues/6

We're going through the repos removing the snyk dependency (check the issue of details). We can only assume this had the dependency once because we have a config file. Just the dependency is not listed in `package.json` anymore.

So, this change is a little different from the other repos in that we are just deleting the snyk exception file.