Persist data for sroc bill run CSV export (#1722)
* Persist data for sroc bill run CSV export

https://eaflood.atlassian.net/browse/WATER-3586

Most of the changes needed for exporting bill run CSV files are done in the UI. However while working on this we have spotted some issues where data is not being correctly persisted in the service.

* Persist `calcWinterDiscountFactor` in db

While working on the export we realised that `calcWinterDiscountFactor` was not being persisted in the db. We therefore update the transaction mapping to ensure this is done.