Add Delete All Billing Data environment variable (#1728)
* Add Delete All Billing Data environment variable
For integration test purposes we need to be able to use the Delete All Bill Runs function in the `preprod` environment.

The logic was previously "Only enable this for specific values of `NODE_ENV`". We change this to use its own flag, `ENABLE_DELETE_ALL_BILLING_DATA_BUTTON`, and as a failsafe we also check that the environment is not `production`.

* Update logic

* Rename flag

We realised that `ENABLE_DELETE_ALL_BILLING_DATA_BUTTON`, the name we took from the UI, wasn't relevant to the service! So we change it to `ENABLE_DELETE_ALL_BILLING_DATA_FEATURE`.