Postal address issue - water abstraction alerts
https://eaflood.atlassian.net/browse/WATER-3717

When creating a water abstraction alert, the alert will be sent by letter to the licence holder if no water abstraction alert contact has been added and opted in to receive email alerts.

As part of the 'Create a water abstraction alert' process the postal address the alert is sent to is being cut down, removing lines 1 and 2 from the address.