Fix address concatenation (#1733)
https://eaflood.atlassian.net/browse/WATER-3726

For some of the address when all four address fields in the database are used, we concatenate name and first line of address.

Expected result: In the instance where all address fields are used, We can concatenate any  address fields other than  name, first line of address and postcode.

* Amended approach
We're splitting the address into a fixed-format head and tail, with the body in the middle which we can amend in order to fit the allowed number of lines

* Fix failing unit tests
A couple of other modules make use of the Notify address formatting. These tests started failing as they expected the address in the old format. We have agreed that we can use the same format in all places, so we update the unit tests to pass with the new format.

Co-authored-by: Stuart Adair <43574728+StuAA78@users.noreply.github.com>