Sroc upload -- allow public water only for water undertakers (#1744)
* Add `testNotWaterUndertaker` validation
We create and apply the validation to check if the public water field is `Y` and if so, is the licence a water undertaker

* Small validation refactor

We noticed a lot of `if...else` statements in the validators which weren't necessary as each bit ended with a `return`. We do a quick refactor pass to eliminate these unnecessary `else` statements to make it a little clearer to read.