Config 2nd service to run backk ground jobs
The service is responsible for handling multiple background process as well as all requests from the other apps. It is the main point of contact for anything to do with the water schema for example.

This means if it needs to run any intensive processes all those other things slow down and in some cases errors are seen. For example, the charge versions import seems to be causing the service to use 100% of the available resources allocated to it, which was leading to other requests erroring.

We want to look to see if, like the UI, we can create a second instance of the service using pm2 to manage it. If we can, we can then start to look at using the 'background' instance to run the more intensive jobs and the normal instance handling requests from the other apps.

So, in this initial change we are just going to look at the pm2 config and using a different startup file.