SRoC - increase the timeout for the import/ upload process
https://eaflood.atlassian.net/browse/WATER-3742

During BAT It was noted that some uploads were failing and the system was erroring in various places.

From reviewing the logs we know one of the reasons was that BullMQ was failing to ‘lock jobs' in a sufficient amount of time. Digging into the BullMQ issues on its repo we found a primary cause of this is low system resources. We could see that the API service which runs the job was hogging 100% of CPU time, which was the cause of the errors being seen i.e. low system resources.