Refactor CSV upload for performance (#1757)
* Refactor CSV upload loops

https://eaflood.atlassian.net/browse/WATER-3745

We noticed that the CSV upload function iterates over arrays by using a `while` loop and `shift()` to get each element to be handled. In our testing we found that replacing these with `for...of` loops was more performant; a ~700 line file previously took 6 mins to import locally with the process using ~1GB memory, whereas changing to a `for...of` loop took 3.5 mins to import using ~150MB memory.