Remove case conversion
When validating, we are constantly converting from camel case to snake case and back using lodash helper functions. This is so we can eg. write the contents of `csvFields` in camel case but they will still align with the snake case column names in the CSV. If we rewrite `csvFields` to use snake case then we don't need to constantly do this conversion.