Refactor `skip` and `allow` in `csvFields`
The `csvFields.js` file which defines the validators for each field allows us to generically test for allowed values: `skip` allows us to skip validation if the specified validator fails, while `allow` lets us specify values and we skip all validation if the field contains one of them.

In practice, these were only being used to skip validation if another field is empty in two very specific cases; and skip validation if the current field is empty.

We therefore remove the former completely and add additional checks to our main validation loop for these two cases (which should improve performance by no longer needing to do a lookup on the desired field); and change the latter for an `allowedEmpty` flag (which should improve performance by not trying to match a value against an array).