Apply tests concurrently with `Promise.all`
As written, we were checking each row a field at a time, and each field a validator at a time. We do a small refactor to instead check all fields in a row concurrently, and all validators for each field concurrently. We expect to see a performance increase from this.