Further CSV upload refactoring (#1758)
* Further CSV upload refactoring

https://eaflood.atlassian.net/browse/WATER-3745

We are trying further optimisations to see if we can make the CSV upload even more performant. The main changes are:

- Rewritten the field names in `csvFields` from camel case to snake case to avoid unnecessary case conversion
- Remove unneeded `testNotBlank` validation, where we were first validating the field isn't blank and then validating for specific values; we now only test for the values as this will reject if the field is empty
- Specify exact columns where we can instead of looking up the column, as the CSV layout will not be changing
- Rework `testNumber` to use native JS instead of Joi, for a performance boost
- Refactor `csvFields` so generic `skip` and `allow` properties (which were only being used in very specific circumstances) are now tested for more specifically
- We now test all fields within a row, and all validators for a field, concurrently using `Promise.all`

* Remove case conversion

When validating, we are constantly converting from camel case to snake case and back using lodash helper functions. This is so we can eg. write the contents of `csvFields` in camel case but they will still align with the snake case column names in the CSV. If we rewrite `csvFields` to use snake case then we don't need to constantly do this conversion.

* Remove unneeded `testNotBlank`

If we are using `testAcceptedTerm` to validate that a field contains one of a choice of text items, we don't also need to test whether it's blank as this would cause `testAcceptedTerm` to fail. We therefore remove `testAcceptedTerm` in these scenarios

* Specify exact column for `licence_number`

* Remove lodash `get`

In two of our validators we use lodash's `get` function. We rework them to use native JS instead.

* Remove optional chaining

It turns out optional chaining kills our performance, so we replace it with a guard clause

* Amend `testNumber` to use native JS

The `testNumber` validator was using Joi to validate the value. In our benchmarking we found that this was considerably slower than native JS -- not so great when we need to validate hundreds of values as quickly as possible. We therefore rewrite the validator in native JS.

* Refactor `skip` and `allow` in `csvFields`

The `csvFields.js` file which defines the validators for each field allows us to generically test for allowed values: `skip` allows us to skip validation if the specified validator fails, while `allow` lets us specify values and we skip all validation if the field contains one of them.

In practice, these were only being used to skip validation if another field is empty in two very specific cases; and skip validation if the current field is empty.

We therefore remove the former completely and add additional checks to our main validation loop for these two cases (which should improve performance by no longer needing to do a lookup on the desired field); and change the latter for an `allowedEmpty` flag (which should improve performance by not trying to match a value against an array).

* Apply tests concurrently with `Promise.all`

As written, we were checking each row a field at a time, and each field a validator at a time. We do a small refactor to instead check all fields in a row concurrently, and all validators for each field concurrently. We expect to see a performance increase from this.

* Remove empty lines

* Remove commented code

* Remove lodash `snakeCase`

We no longer need to convert headings to snake case in order to validate that the correct columns are present