Identify potential optimisation in charge version creation
Creating a charge version is something we do in several places. Looking at the existing code, it's pulling down all the existing charge versions for a licence from a db just to update their status and dates. We feel we can optimise this by going direct to the db so we add some `TODO` comments to flag these up for later review