Refactor `update-charge-information-save` loop (#1760)
* Refactor `update-charge-information-save` loop

https://eaflood.atlassian.net/browse/WATER-3745

As with other parts of the charge versions upload code, we switch out a `do...while` loop with `shift()` for ` regular `for...of` loop to help optimise it.

* Change `do...while`/`shift` loop to `for...of`

* Identify potential optimisation in charge version creation

Creating a charge version is something we do in several places. Looking at the existing code, it's pulling down all the existing charge versions for a licence from a db just to update their status and dates. We feel we can optimise this by going direct to the db so we add some `TODO` comments to flag these up for later review