Move towards a WorkerManager strategy
We were initially going to more forward with the existing `QueueManager` running in the foreground instance pinging across to the background instance what workers to create.

There was an initial concern about that adding a startup constraint (background needs to be up before foreground else POST requests would fail) but we were gonna return to it later. Then we spotted the QueueManager was keeping track of the worker instances so that it could gracefully stop them when the service instance is stopped.

That was going to make the logic in our background tasks controller more complex, because now we needed to think about trying to do the same thing. With the startup constraint as well, we decided to ditch using an endpoint and instead split the responsibility.

With this change we add a new `WorkerManager`. All the logic for dealing with workers was migrated from QueueManager as well as shifting responsibility for stopping the workers when the service instance is stopped.

We've left our background tasks controller because we think we may want to add endpoints in the future, if only for debug purposes. If nothing else it gives us somewhere to put our `/status` endpoint which never felt right just hardcoded into the routes file.

Finally, at this time we can't see a need for the foreground to talk to the background so we say bye-bye to the background connecter.