Migrate QueueScheduler / Tidying the code
After banging our heads trying to understand why when killing the foreground app all the works stopped in the background app we spotted the problem.

Repeating jobs depend on BullMQ's delay functionality and that only works for queues with a QueueScheduler. The QueueManager was creating these and therefore when the foreground app died so did the QueueSchedulers.

Migrating the instantiation of these to the WorkerManager meant repeatable jobs worked again even when the foreground app was closed.

After that was just some final tweaks, mainly around when the server instances stop.