Add remaining workers to be registered
What we are registering is based on the output we see when the foreground service starts up; QueueManager logs all the queues it creates, which previously would have also been all the workers.

We took that list and used it to create our `RegisterWorkersService`. This specific change is us adding the remaining workers and shifting everything into alphabetical order.