Remove `config` folder (#1772)
* Remove `config` folder

We realise that we have a top-level folder `config` which contains one single folder, `task-config`. This relates to `scripts/migrate-task-config`. To avoid any confusion when pulling in the top-level `config.js` file, we therefore move this folder into `scripts` and keep the `migrate-task-config` code in one place.

* Move `config/task-config`

We move the folder `config/task-config` to `scripts`, renaming it to `migrate-task-config` to match the calling script