Correctly set `NODE_ENV` (#1771)
* Correctly set `NODE_ENV`

https://github.com/DEFRA/water-abstraction-team/issues/43

We are currently using `NODE_ENV` incorrectly; we set it to be the name of the specific environment (eg. `local`, `qa` etc.) when it should be more like the _type_ of environment (eg. is it a dev environment, a production environment, etc.)

This means that tools like pm2 aren't correctly picking up that we are running in a dev-like or prod-like environment and configuring accordingly.

We therefore change how we use `NODE_ENV` to be the type of environment, with a separate `ENVIRONMENT` env var to set the actual environment (eg. `tst`, `pre` etc.)

* Streamline environment env vars

We change most of our environment-related env vars to use a new `ENVIRONMENT` var, which specifies the exact environment (eg. dev, test etc.). And where needed, we retain `NODE_ENV` to specify the type of environment (ie. test or production)

* Remove `isLocal`

We only use `isLocal` to configure the number of concurrent jobs on a local machine. We remove this so that the number of jobs is the same in all environments

* Remove `isTest`

Again, we remove a config variable that only existed to configure things differently in a test environment

* Update `.env.example`