Increase V8's old memory section via pm2 config
`--max-old-space-size` is an argument you can set when calling node that it will pass on to tje underlying V8 engine. From the node documents

> Sets the max memory size of V8's old memory section. As memory consumption approaches the limit, V8 will spend more time on garbage collection in an effort to free unused memory.
>
> On a machine with 2 GiB of memory, consider setting this to 1536 (1.5 GiB) to leave some memory for other uses and avoid swapping.
https://nodejs.org/api/cli.html#useful-v8-options

The default is 2GB which can be seen by checking `heap_size_limit` after doing the following

```
node

> v8.getHeapStatistics()
{
  ....
  heap_size_limit: 2197815296,
}
```
https://stackoverflow.com/a/63495296/6117745

The essence of this change is to improve performance. The background instance is where all our work is done and where we often have to hold a large amount of data in memory. The theory is the more memory you've got to hold stuff, the less the garbage collector has to be called in to ensure you don't run out of memory.

We're only setting this when running in `production` mode, which will only be when running in our AWS environments. Note [PR #1771](https://github.com/DEFRA/water-abstraction-service/pull/1771) was crticial to us being able to do this.