Fix issue with worker onComplete
We thought all was working but actually we'd broken something. The original QueueManager would pass a reference to itself to the worker `onComplete` handlers. The intent is they can then use it to add the next job onto the queue, if relevant.

When we wrote WorkManager based on this code we kept the `, this)` param call. But that meant we were passing the workerManager instance to the worker onComplete handlers. It has no `add()` method so the app was blowing up at this point.

With this change we make use of the existing QueueManager plugin to decorate the server instance. With that in place we can pass it through to RegisterWorkersService, which can then pass it through to WorkerManager.register(), which can finally pass it onto the worker onComplete handlers.

And lo, normal service resumes!