Register jobContainer with QueueManager
So, we're still getting an error. The issue now is the worker `onComplete` handler will call `QueueManager.add()`. In `add()` the QueueManager tries to find the BullMQ `Queue` instance to use to add the job to. But because we've not been calling `QueueManager.register()` with our jobContainers it finds nothing so explodes.

It looks like to support workers completing and being able to add new jobs to queues, as we register job containers in WorkerManager we also need to let QueueManager know about them.