Do a 180 on implementing WorkerManager
> This is a big commit with lots of changes. Think of it as a spike within a spike. We intend to take these changes and implement some of them in a separate PR so we can make the change smaller, and the commit history clearer.

Based on issues found in initial testing we tracked down that the background instance needs to have an instantiated instance of `QueueManager`, with all queues registered. This is because of the nature of a lot of the jobs. In their `onComplete()` handlers they'll add new jobs and expect access to `QueueManager` to do so.

When we stepped back and looked at `WorkerManager` and `QueueManager` side-by-side we realised there was little need for `WorkerManager` to exist. Especially, in light of the trick we learnt to only run certain code if we are the background instance (using the `name` env var pm2 provides).

With that gone the index files also start looking quite similar, or so we thought initially. We updated index-background.js to just add the plugins like index.js and things did work. But then we spotted that a number of the `registerSubscriber()` methods also add jobs to queues. We only want to be adding jobs in the foreground instance with the background justing doing the work.

Plus we kinda liked `RegisterWorkersService`. It was one place you could go in the system to see all the BullMQ 'tasks' we have.

So, after making initial changes to `QueueManager` to only create workers if it's the background instance being used, we made some more changes. `RegisterWorkersService` became `JobRegistrationService` and moved to the `src/lib/message-queue-v2` folder. We created `StartUpJobsService` then looked at all the register subscriber calls being made. Out of those we tracked down the jobs being added and copied them to `StartUpJobsService`.

In both startup index.js files we now use `JobRegistrationService` to register things with the `QueueManager`. And the foreground index.js we also use `StartUpJobsService` to add the same jobs the registerSubscriber calls were adding.

This means we can actually go through the modules and delete/strip out all the 'plugin' stuff. But we decided to stop here and instead do most of these changes in a new PR. They will then become preparatory work and we can get back to this PR focusing on changes specific to supporting the background instance.