Refactor message-queue to centralise registration
https://eaflood.atlassian.net/browse/WATER-3745

> Part of a body of work to support 2 instances of the service running on our AWS servers; one focused on API requests and adding initial BullMQ jobs to queues. The other where the BullMQ workers are created, sat in the background getting on with processing those jobs. Splitting the responsibilities fixes the issue of the service failing to respond to API requests because it's busy processing a job.

The module is called `message-queue-v2` though its content is all about `QueueManager` (QM). QM is a service class currently responsible for creating the BullMQ [queues](https://docs.bullmq.io/guide/queues), and the [workers](https://docs.bullmq.io/guide/workers) and, if required, the [queue schedulers](https://docs.bullmq.io/guide/queuescheduler) associated with them.

It also has helper methods for adding jobs (actually just a JSON object) to queues, and stopping all the workers during shut down.

It knows what to create through a 'registration' process. During startup, a plugin decorates both the Hapi [server](https://hapi.dev/api/?v=20.2.2#server) and [request](https://hapi.dev/api/?v=20.2.2#request) objects with a singleton instance of the QM. Then other modules _register as subscribers_. By this we mean they have there own plugins which when registered with Hapi trigger the 'registration' process. What the service refers to as a 'job' is passed to the QM's `register()` function by the modules.

All jobs follow a pattern of what they are expected to export, including a **Job Name** and an `onFailure()` handler. The QM uses this information to instantiate the corresponding **Queue**, **Worker**, and **QueueScheduler**, and keeps a reference to them.

Some of the modules also add initial jobs to the queues. Often these are repeatable, i.e. once the job is processed create another one to be handled at a specific time, or in a certain amount of seconds.

The primary reason for this refactoring is we want to be able to split out which instance creates the workers. Both need access to the queues because it's a common pattern that when a job completes, it's first action is to add the next job to another queue. But only the background instance needs to create the workers.

With that in mind we need more control other how things are registered. Specficially, we only want those initial startup jobs being added in the foreground instance.

Along the way we'll do some housekeeping, and sort out the naming debacle. This should then set us up nicely for sorting out `index-background.js` and getting the workers running only there.