Move queue cleaning to QueueManager
The main use of `QueueManager.deleteKeysByPattern()` is just prior to registering a job in `register-subscribers()`, it will be called to ensure the new queue starts empty.

We don't have the context for why this is good for some but not others. But we feel this is something that should be done as part of registering the job with the QueueManager.