Fix tests not using test redis DB
https://github.com/DEFRA/water-abstraction-team/issues/4

As part of [Apply consistent CI workflow](https://github.com/DEFRA/water-abstraction-service/pull/1710) we inadvertently broke the `isPermitsTestDatabase` property in `config.js`. This is used to determine which db to use in Redis.

You may not be aware, but you can setup [multiple DB's in Redis](https://redis.io/commands/select/). You simply need to give it a number and Redis will create a DB for it. By default it uses 0 but you can select something else in the connection and Redis will create a DB there.

We take advantage of this feature to

- use different db's for different apps (though it's not consistent)
- use a different db when running the unit tests (again, it's not consistent)

Recently, we've hit an issue when trying to shut down the app it is stalling. The job causing the issue is `notify.send`. It keeps failing because it has jobs with notification ID's but they don't exist in the PostgreSQL db. After much digging, we found the issue only happened after running tests. Turns out because we send real notificiations via Notify in the tests (yes, [we know!](https://github.com/DEFRA/water-abstraction-team/issues/35)) we were creating test jobs in the normal redis database with ID's that didn't match what was in the normal PostgreSQL db. They should have been going into the test redis DB.

The root cause was the `isPermitsTestDatabase` property in `config.js`. It was setting to true only if the database url provided included the DB name 'permits-test'. We changed that though to match what we call it locally and in our CI pipeline ('wabs_test'). So, now it's always coming up as `false` causing jobs added whilst running the unit tests to appear in the DB when just running it locally.

This change fixes the issue.